pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-yield-optimizer"
version = "0.1.0"
description = "DeFi yield optimization with quantum computing"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.12.0",
]

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["data_pipeline"]
